requirements.txt
fastapi
uvicorn
pinecone
groq
python-dotenv
pydantic
aiofiles
pypdf2
httpx
beautifulsoup4 